feature(Administration): Secteur Tutoriel
- Listing des tutoriels
- Création d'un tutoriel
- Edition d'un tutoriel
- Suppression d'un tutoriel

BREAKING CHANGE: [Admin][Tutoriel] - Gestion des Vidéos

Close #14

+++ resources/js/admin/objet/objet/index.js
@@ -121,11 +121,17 @@ function loadTable() {
                 autoHide: false,
                 textAlign: 'right',
                 template: function (row) {
-                    return `
-                    <a href="/administrator/objet/objet/${row.id}" class="btn btn-icon btn-sm btn-default"><i class="la la-eye"></i> </a>
-                    <a href="/administrator/objet/objet/${row.id}/edit" class="btn btn-icon btn-sm btn-info"><i class="la la-edit"></i> </a>
-                    <a href="/administrator/objet/objet/${row.id}/delete" class="btn btn-icon btn-sm btn-danger"><i class="la la-trash-o"></i> </a>
+                    if (row.published === 0) {
+                        return `
+                        <a href="/administrator/objet/objet/${row.id}" class="btn btn-icon btn-sm btn-default"><i class="la la-eye"></i> </a>
+                        <a href="/administrator/objet/objet/${row.id}/edit" class="btn btn-icon btn-sm btn-info"><i class="la la-edit"></i> </a>
+                        <a href="/administrator/objet/objet/${row.id}/delete" class="btn btn-icon btn-sm btn-danger"><i class="la la-trash-o"></i> </a>
                     `;
+                    } else {
+                        return `
+                        <a href="/administrator/objet/objet/${row.id}" class="btn btn-icon btn-sm btn-default"><i class="la la-eye"></i> </a>
+                    `;
+                    }
                 }
             }
         ],
